Changed group ID to org.wso2.performance.common
