+++ components/jtl-splitter/src/main/java/org/wso2/performance/common/jtl/splitter/JTLSplitter.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.wso2.performance.jtl.splitter;
+package org.wso2.performance.common.jtl.splitter;
 
 import com.beust.jcommander.JCommander;
 import com.beust.jcommander.Parameter;
