+++ components/netty-http-echo-service/src/main/java/org/wso2/performance/common/netty/echo/EchoHttpServer.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.wso2.performance.netty.echo;
+package org.wso2.performance.common.netty.echo;
 
 import com.beust.jcommander.JCommander;
 import com.beust.jcommander.Parameter;
