+++ components/netty-http-echo-service/src/main/java/org/wso2/performance/common/netty/echo/EchoHttpServerHandler.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.wso2.performance.netty.echo;
+package org.wso2.performance.common.netty.echo;
 
 import io.netty.channel.ChannelFutureListener;
 import io.netty.channel.ChannelHandler.Sharable;
