+++ components/payload-generator/pom.xml
@@ -17,7 +17,7 @@
 <project xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xmlns="http://maven.apache.org/POM/4.0.0"
          xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
     <parent>
-        <groupId>org.wso2.performance</groupId>
+        <groupId>org.wso2.performance.common</groupId>
         <artifactId>performance-common</artifactId>
         <version>0.1.0-SNAPSHOT</version>
         <relativePath>../../pom.xml</relativePath>
@@ -45,7 +45,7 @@
     </build>
 
     <properties>
-        <main.class>org.wso2.performance.payload.generator.PayloadGenerator</main.class>
+        <main.class>org.wso2.performance.common.payload.generator.PayloadGenerator</main.class>
     </properties>
 
 </project>
