+++ components/payload-generator/src/main/java/org/wso2/performance/common/payload/generator/PayloadGenerator.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.wso2.performance.payload.generator;
+package org.wso2.performance.common.payload.generator;
 
 import com.beust.jcommander.JCommander;
 import com.beust.jcommander.Parameter;
