+++ findbugs-exclude.xml
@@ -17,17 +17,17 @@
 <FindBugsFilter>
     <Match>
         <Bug pattern="NP_NULL_ON_SOME_PATH_FROM_RETURN_VALUE"/>
-        <Class name="org.wso2.performance.jtl.splitter.JTLFileValidator"/>
+        <Class name="org.wso2.performance.common.jtl.splitter.JTLFileValidator"/>
         <Method name="validate"/>
     </Match>
     <Match>
         <Bug pattern="NP_NULL_ON_SOME_PATH_FROM_RETURN_VALUE"/>
-        <Class name="org.wso2.performance.jtl.splitter.JTLSplitter"/>
+        <Class name="org.wso2.performance.common.jtl.splitter.JTLSplitter"/>
         <Method name="splitJTL"/>
     </Match>
     <Match>
         <Bug pattern="DM_DEFAULT_ENCODING"/>
-        <Class name="org.wso2.performance.jtl.splitter.JTLSplitter"/>
+        <Class name="org.wso2.performance.common.jtl.splitter.JTLSplitter"/>
         <Method name="splitJTL"/>
     </Match>
 </FindBugsFilter>
