+++ pom.xml
@@ -23,7 +23,7 @@
     </parent>
 
     <modelVersion>4.0.0</modelVersion>
-    <groupId>org.wso2.performance</groupId>
+    <groupId>org.wso2.performance.common</groupId>
     <artifactId>performance-common</artifactId>
     <version>0.1.0-SNAPSHOT</version>
     <packaging>pom</packaging>
@@ -103,17 +103,17 @@
             </dependency>
             <!-- Distribution Dependencies - Start -->
             <dependency>
-                <groupId>org.wso2.performance</groupId>
+                <groupId>org.wso2.performance.common</groupId>
                 <artifactId>netty-http-echo-service</artifactId>
                 <version>${performance.common.version}</version>
             </dependency>
             <dependency>
-                <groupId>org.wso2.performance</groupId>
+                <groupId>org.wso2.performance.common</groupId>
                 <artifactId>payload-generator</artifactId>
                 <version>${performance.common.version}</version>
             </dependency>
             <dependency>
-                <groupId>org.wso2.performance</groupId>
+                <groupId>org.wso2.performance.common</groupId>
                 <artifactId>jtl-splitter</artifactId>
                 <version>${performance.common.version}</version>
             </dependency>
